change to use a scala version sensitive branch
